sql functions receive patid as bigint

+++ sql/functions/createHesApptVisitId.sql
@@ -4,7 +4,7 @@ Result is a integer of length 18, consisting of attendkey with (part of) patient
 NOTE: This id is not guaranteed to be unique for each attendkey/patientid combination.
 However, incorporating the whole patientId is not possible (bigint max 19 places) and this is very unlikely to give clashes.
 */
-CREATE OR REPLACE FUNCTION createHesApptVisitId(attendkey TEXT, patientId INTEGER)
+CREATE OR REPLACE FUNCTION createHesApptVisitId(attendkey TEXT, patientId BIGINT)
   RETURNS BIGINT AS
 $$
 DECLARE
